Refactor: 헤더 로고, 오른쪽 아이콘  컴포넌트 분리
