+++ src/layout/Header/RightIcons/ThemeMode/index.jsx
@@ -1,5 +1,5 @@
 import { useSelector, useDispatch } from 'react-redux';
-import { darkMode, lightMode } from '../../../store/modules/header';
+import { darkMode, lightMode } from '../../../../store/modules/header';
 import { HiMoon } from 'react-icons/hi';
 import { BsFillSunFill } from 'react-icons/bs';
 import styled from 'styled-components';
