+++ src/layout/Header/RightIcons/ToggleMenuList/index.jsx
@@ -63,6 +63,9 @@ const ToggleMenuList = ({ toggleMenuRef, setIsToggleOpen }) => {
 export default ToggleMenuList;
 
 const ToggleMenuListContainer = styled.div`
+  position: absolute;
+  top: 3.5rem;
+  right: 0;
   width: 12rem;
   margin-top: 0.3rem;
   margin-left: auto;
